EternalAI.py/interact/pyproject.toml
[project]
name = "interact"
version = "0.1.0"
description = ""
authors = [
    { name = "0xgen0000", email = "52655944+0xgen0000@users.noreply.github.com" }
]
readme = "README.md"
requires-python = ">=3.8,<4"
dependencies = [
    "web3 (>=6.0,<7.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "uniswap-python (>=0.7.1,<0.8.0)",
    "simplejson (>=3.20.1,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"

